wheel_requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.39.0
celery==3.1.23
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-axes==1.6.0
django-cors-headers==1.1.0
django-countries==4.0
django-formtools==1.0
django-otp==0.3.5
django-phonenumber-field==0.7.2
django-rest-auth==0.8.1
django-rest-swagger==0.3.5
django-ses-backend==0.1.1
django-storages==1.5.0
django-timezone-field==1.3
django-two-factor-auth==1.3.1
django-user-sessions==1.3.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.4.5
httplib2==0.9.2
kombu==3.0.35
Markdown==2.6.6
oauthlib==1.0.3
phonenumbers==7.3.0
Pillow==3.2.0
psycopg2==2.6.2
PyJWT==1.4.0
pysha3==0.3
PySocks==1.5.6
python3-openid==3.0.10
pytz==2016.3
PyYAML==3.11
qrcode==4.0.4
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
twilio==5.4.0
web3==0.1.9